+++ test/helpers/main.js
@@ -1,0 +1,37 @@
+import path from 'node:path';
+import process from 'node:process';
+import {setTimeout} from 'node:timers/promises';
+import {fileURLToPath} from 'node:url';
+import {multibyteFirstHalf, multibyteSecondHalf} from './arguments.js';
+
+export const isLinux = process.platform === 'linux';
+export const isWindows = process.platform === 'win32';
+
+export const FIXTURES_URL = new URL('../fixtures/', import.meta.url);
+export const fixturesPath = fileURLToPath(FIXTURES_URL);
+
+export const nodeDirectory = path.dirname(process.execPath);
+
+export const earlyErrorOptions = {detached: 'true'};
+
+// TODO: replace with Array.fromAsync() after dropping support for Node <22.0.0
+export const arrayFromAsync = async asyncIterable => {
+	const chunks = [];
+	for await (const chunk of asyncIterable) {
+		chunks.push(chunk);
+	}
+
+	return chunks;
+};
+
+export const destroySubprocessStream = async ({nodeChildProcess}, error, streamName) => {
+	const subprocess = await nodeChildProcess;
+	subprocess[streamName].destroy(error);
+};
+
+export const writeMultibyte = async promise => {
+	const {stdin} = await promise.nodeChildProcess;
+	stdin.write(multibyteFirstHalf);
+	await setTimeout(1e2);
+	stdin.end(multibyteSecondHalf);
+};
